refactor(onboard): split Ollama Linux upgrade helpers
